requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.14.0
idna==3.11
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.5
sqlparse==0.5.5
stripe==14.1.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
